Fixing data type for grades

+++ program_functions.cpp
@@ -306,7 +306,7 @@ void withdraw_from_course(string sid, string cid, vector<Student> &slist, vector
 
 //10. update student grade that takes a student id, course id, and grade to update the student grade in a given course. note that parameters needed
 // might be more than the three specified
-void update_grade(string sid, string cid, int new_grade, vector<Student> &slist, vector<Course> &clist)
+void update_grade(string sid, string cid, double new_grade, vector<Student> &slist, vector<Course> &clist)
 {
 
     Student *current_s = find_student(sid, slist);
